Cleanup, organization, and documentation
