+++ lib/reviewer/keywords/git/staged.rb
@@ -17,6 +17,12 @@ module Reviewer
           stdout.strip.empty? ? [] : stdout.split("\n")
         end
 
+        # Gets the list of staged files
+        #
+        # @example Get the list of files
+        #   staged.list #=> ['/Code/example.rb', '/Code/run.rb']
+        #
+        # @return [Array<String>] the array of staged filenames as strings
         def list
           @stdout, @stderr, @status = Open3.capture3(command)
           @exit_status = @status.exitstatus.to_i
@@ -24,10 +30,20 @@ module Reviewer
           @status.success? ? to_a : raise_command_line_error
         end
 
+        # Convenience method for retrieving the list of staged files since there's no parameters
+        #   for an initializer.
+        #
+        # @example Get the list of files
+        #   Reviewer::Keywords::Git::Staged.list #=> ['/Code/example.rb', '/Code/run.rb']
+        #
+        # @return [Array<String>] the array of staged filenames as strings
         def self.list
           new.list
         end
 
+        # Assembles the pieces of the command that gets the list of staged files
+        #
+        # @return [String] the full command to run to retrieve the list of staged files
         def command
           command_parts.join(' ')
         end
